cleaned up planner, verified working
